requirements-api.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
gtts==2.5.4
pydub==0.25.1
pyyaml==6.0.2
requests==2.32.3
edge-tts==7.2.8
kokoro-onnx==0.5.0
soundfile==0.12.1
piper-tts==1.4.2
numpy
